Extract custom field `user_id` from JWT token (Slack external ID) for consistent user identification
The `user_id` field is extracted from the decoded JWT token. This value corresponds to the Slack external ID and is used to identify active users in our database. Centralizing this extraction allows us to access the user ID consistently throughout the application.

Remove `userClerkId` from context as we now have the `user_id`