Feat/extract userid from decoded token (#45)
* Extract custom field `user_id` from JWT token (Slack external ID) for consistent user identification

The `user_id` field is extracted from the decoded JWT token. This value corresponds to the Slack external ID and is used to identify active users in our database. Centralizing this extraction allows us to access the user ID consistently throughout the application.

Remove `userClerkId` from context as we now have the `user_id`

* Improve distributor repository and service to:
- Take ID as string instead of uint
- Use new method `findOneWithField`  that allow fields selections and model preloads

* Update distributor and workspace handlers to use `user_id` from context instead of `clerkUserId`

* Improve workspace creation method to not find distributor before creating the workspace, and just use the user_id from context